+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Changelog
-All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [7.0.1] - 2025-01-26
+### Changed
+- Improved some type annotations and simplified internal test configuration.
+- Switched from Psalm to PHPStan for static analysis.
+
+### Fixed
+- `floatSelectedAsString` option is now `false` for PostgreSQL on PHP 8.4.
+
 
 ## [7.0.0] - 2024-10-29
 ### Added
@@ -342,6 +347,7 @@ return the value of their callback function, making it easier to use data outsid
 - Initial release
 
 
+[7.0.1]: https://github.com/devtheorem/peachy-sql/compare/v7.0.0...v7.0.1
 [7.0.0]: https://github.com/devtheorem/peachy-sql/compare/v6.3.1...v7.0.0
 [6.3.1]: https://github.com/devtheorem/peachy-sql/compare/v6.3.0...v6.3.1
 [6.3.0]: https://github.com/devtheorem/peachy-sql/compare/v6.2.0...v6.3.0